Fixed bad package naming
Moved project into seperate package from root, update file structure, package names and imports.
